Improve accuracy and efficiency of Stirling evaluations for Gamma and Beta. Apply improved Poission evaluation to Gamma distribution too.
